requirements.txt
tk==0.1.0
tkintermapview==1.24
customtkinter==5.0.3
Pillow==9.3.0
requests==2.28.1
psycopg2==2.9.5